# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Harness Studio API
  version: 0.2.0
servers:
  - url: http://localhost:8080
paths:
...
  /api/reports:
    post:
      summary: Create a report job
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Report job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
  /api/reports/{id}:
    get:
      summary: Get report job status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Report job status
  /api/batches/{id}/report:
    post:
      summary: Create and run a batch report job
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "202":
          description: Report generation requested
    get:
      summary: Get latest batch report job
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Latest batch report job
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Envelope:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        statusCode:
          type: integer
        data: {}
...